+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - **Added:** Support for `appearance` prop in `BoolField` in `uniforms-material`. [\#189](https://github.com/vazco/uniforms/issues/189)
 - **Changed:** Use `label` instead of specified props in `ListAddField` in `uniforms-material`.
 - **Changed:** Use `label` instead of specified props in `ListDelField` in `uniforms-material`.
+- **Fixed:** Handling of `defaultValue` in `SimpleSchema2Bridge`. [\#182](https://github.com/vazco/uniforms/issues/182)
 - **Fixed:** `ListAddField` in `uniforms-material`.
 
 ## [v1.10.0](https://github.com/vazco/uniforms/tree/v1.10.0) (2017-02-07)